Move iris_df fixture to conftest.py
