+++ tests/test_associations.py
@@ -7,28 +7,6 @@ from sklearn import datasets
 from dython.nominal import associations
 
 
-@pytest.fixture
-def iris_df():
-    # Use iris dataset as example.
-    # Add one made-up categorical column to create a nom-nom relationship.
-
-    iris = datasets.load_iris()
-
-    target = ['C{}'.format(i) for i in iris.target]
-
-    extra = target[50:]
-    extra.extend(target[:50])
-
-    extra = pd.DataFrame(data=extra, columns=['extra'])
-
-    X = pd.DataFrame(data=iris.data, columns=iris.feature_names)
-    y = pd.DataFrame(data=target, columns=['target'])
-
-    df = pd.concat([X, extra, y], axis=1)
-
-    return df
-
-
 @pytest.fixture(autouse=True)
 def disable_plot(monkeypatch):
     # Patch plt.show to not halt testing flow, by making it not block
